♻️(recrutement-presentation) isolate RecrutementCandidaturesEtapeView to recrutement_detail.py for more clarity
